# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "minotari",
    "description": "",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.2.0"
  },
  "paths": {
    "/accounts/{name}/address": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves the Tari address for a specified account.",
        "description": "Returns the account's Tari address in Base58 format along with the emoji ID\nrepresentation. This address can be shared with others to receive payments.\n\n# Path Parameters\n\n- `name`: The unique account name to query\n\n# Response\n\nReturns an [`AddressResponse`] object containing:\n- The address in Base58 encoding\n- The emoji ID representation\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Response\n\n```json\n{\n  \"address\": \"f4FxMqKAPDMqAjh6hTpCnLKfEu3MmS7NRu2YmKZPvZHc2K\",\n  \"emoji_id\": \"🎉🌟🚀...\"\n}\n```",
        "operationId": "api_get_address",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve address for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account address retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{name}/address_with_payment_id": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Creates a Tari address with an embedded payment ID for a specified account.",
        "description": "Generates a new address that includes a payment ID, which can be used to\nidentify specific transactions or invoices. When someone sends funds to\nthis address, the payment ID will be included in the transaction.\n\n# Path Parameters\n\n- `name`: The unique account name\n\n# Request Body\n\nSee [`CreatePaymentIdAddressRequest`] for the complete request schema.\n\n# Response\n\nReturns an [`AddressWithPaymentIdResponse`] object containing:\n- The address with embedded payment ID in Base58 encoding\n- The emoji ID representation\n- The original payment ID\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n- [`ApiError::InternalServerError`]: Failed to generate the address\n\n# Example Request\n\n```bash\ncurl -X POST http://localhost:8080/accounts/default/address_with_payment_id \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"payment_id\": \"696e766f6963652d3132333435\"}'\n```\n\n# Example Response\n\n```json\n{\n  \"address\": \"f4FxMqKAPDMqAjh6hTpCnLKfEu3MmS7NRu2YmKZPvZHc2K\",\n  \"emoji_id\": \"🎉🌟🚀...\",\n  \"payment_id_hex\": \"696e766f6963652d3132333435\"\n}\n```",
        "operationId": "api_create_address_with_payment_id",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to create address for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentIdAddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Address with payment ID created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressWithPaymentIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{name}/balance": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves the current balance for a specified account.",
        "description": "Returns the account's available balance, pending incoming transactions,\nand locked funds. This endpoint is useful for displaying wallet status\nor checking available funds before initiating a transaction.\n\n# Path Parameters\n\n- `name`: The unique account name to query\n\n# Response\n\nReturns an [`AccountBalance`] object containing:\n- Available (spendable) balance\n- Pending incoming balance\n- Locked balance (reserved for pending transactions)\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Response\n\n```json\n{\n  \"available\": 10000000,\n  \"pending_incoming\": 500000,\n  \"locked\": 200000\n}\n```",
        "operationId": "api_get_balance",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve balance for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountBalance"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{name}/burn": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Burns funds from an account and records a partial burn proof for L2 claiming.",
        "description": "Creates a burn transaction, broadcasts it, and stores the partial proof in the\ndatabase. The daemon's `BurnProofWorker` will automatically fetch the kernel\nmerkle proof once the transaction is confirmed and write the complete\n`CompleteClaimBurnProof` JSON file to the configured `burn_proofs_dir`.\n\n# Path Parameters\n\n- `name`: The account to burn funds from\n\n# Request Body\n\nSee [`BurnFundsRequest`] for the complete schema.\n\n# Response\n\nReturns a [`BurnFundsResponse`] with:\n- `tx_id`: the transaction ID\n- `output_hash`: hex-encoded output hash (use this to identify the proof file)\n\n# Errors\n\n- `400 Bad Request`: invalid `claim_public_key` or `sidechain_deployment_key` hex\n- `404 Not Found`: account does not exist\n- `500`: insufficient funds, broadcast failure, or other internal error\n\n# Example Request\n\n```bash\ncurl -X POST http://localhost:9000/accounts/default/burn \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"amount\": 1000000, \"claim_public_key\": \"a3f9...\"}'\n```",
        "operationId": "api_burn_funds",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to burn from",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BurnFundsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Burn transaction broadcast successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BurnFundsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Burn failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts/{name}/displayed_transactions/by_payref/{payref}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves displayed transactions by payment reference.",
        "description": "Returns displayed transactions that contain the specified payment reference.\nThis searches within the payment references stored in each transaction.\n\n# Path Parameters\n\n- `name`: The unique account name to query\n- `payref`: The payment reference to search for\n\n# Response\n\nReturns a list of [`DisplayedTransaction`] objects that contain the payment reference.\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Request\n\n```bash\ncurl -X GET http://localhost:8080/accounts/default/displayed_transactions/by_payref/my-payment-ref-123 \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n```",
        "operationId": "api_get_displayed_transactions_by_payref",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve transactions from",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "payref",
            "in": "path",
            "description": "Payment reference to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Displayed transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayedTransaction"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressResponse": {
        "type": "object",
        "description": "API response type for an account address.\n\nContains the Tari address in Base58 format along with the emoji ID representation.\n\n# JSON Example\n\n```json\n{\n  \"address\": \"f4FxMqKAPDMqAjh6hTpCnLKfEu3MmS7NRu2YmKZPvZHc2K\",\n  \"emoji_id\": \"🎉🌟🚀...\"\n}\n```",
        "required": [
          "address",
          "emoji_id"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "The Tari address in Base58 encoding"
          },
          "emoji_id": {
            "type": "string",
            "description": "The emoji representation of the address"
          }
        }
      },
...
      "CreatePaymentIdAddressRequest": {
        "type": "object",
        "description": "Request body for creating an address with a payment ID.\n\n# JSON Example\n\n```json\n{\n  \"payment_id_hex\": \"696e766f6963652d3132333435\"\n}\n```",
        "required": [
          "payment_id_hex"
        ],
        "properties": {
          "payment_id_hex": {
            "type": "string",
            "description": "The payment ID to embed in the address, hex encoded.\n\nThis should be a hex-encoded byte string (e.g., \"696e766f6963652d3132333435\" for \"invoice-12345\").\nMaximum length is 256 bytes when decoded."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Wallet API token, sent as 'X-API-Key: <token>'"
      },
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Wallet API token, sent as 'Authorization: Bearer <token>'"
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    },
    {
      "api_key": []
    }
  ],
...
}